refactor ConfigDialog
